[StyleCop] Fix all the warnings in YamlParser - Create new files for the classes: Dictionary, List, NestedRegex, ParamsRegex and SimpleRegex - Add and remove spacing when needed
